feat: Pause and resume Scene (#381)
* refactor/scene: simplifying accesor usage

* feat/refactor/debug: debugging camera movement to debugger, speed, ctrl

also add use the main camera as movement target

* cleanup/octree: redundant branch, readability

* fix: casting data type

fix: data widen casting

uint64 casting

* fix: missing HRESULT check

* fix/imgui: different add component dialog stacks each other

* fix/imgui: render component shows nothing for material

* fix/mouseManager: returning local variable as reference

* chore: justification of type casting

* fix/object: referencing values

+++ Engine/egRayTracer.cpp
@@ -64,7 +64,7 @@ namespace Engine::Manager::Graphics
 
     cmd->SoftReset();
 
-    m_light_buffer_data_.SetData(cmd->GetList4(), m_light_buffers_.size(), m_light_buffers_.data());
+    m_light_buffer_data_.SetData(cmd->GetList4(), static_cast<UINT>(m_light_buffers_.size()), m_light_buffers_.data());
 
     RenderPass(cmd->GetList4(), nullptr);
 
@@ -105,7 +105,7 @@ namespace Engine::Manager::Graphics
   {
     if (const auto& scene = GetSceneManager().GetActiveScene().lock())
     {
-      UINT32 total_item_count = 0;
+      UINT64 total_item_count = 0;
 
       // Scrap the BLAS.
       std::map<WeakMaterial, std::vector<SBs::InstanceSB>> target_instances;
